% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.r
\name{dijkstra}
\alias{dijkstra}
\title{determine the shortest path to all the nodes, which are included in the network}
\usage{
dijkstra(wiki_graph, initial_node)
}
\arguments{
\item{wiki_graph}{a data frame, which contains three parameters: the connection path between two nodes and weight function}

\item{initial_node}{from which node, the trip begins}
}
\value{
the expected shortest journey, towards each and every node, through the entire network
}
\description{
determine the shortest path to all the nodes, which are included in the network
}
