% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the largest common divisor}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer}

\item{b}{An integer}
}
\value{
The largest common divisor of \code{a} and \code{b}
}
\description{
Find the largest common divisor
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
