% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factorial.R
\name{factorial_function}
\alias{factorial_function}
\title{Calculate Factorial}
\usage{
factorial_function(n)
}
\arguments{
\item{n}{A non-negative integer.}
}
\value{
The factorial of the input number.
}
\description{
This function calculates the factorial of a non-negative integer.
}
