% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say Hello}
\usage{
hello(name)
}
\arguments{
\item{name}{A character string with the user's name.}
}
\value{
A greeting message.
}
\description{
This function greets the user by name.
}
