\name{ourpackage-package}
\alias{ourpackage-package}
\alias{ourpackage}
\docType{package}
\title{
\packageTitle{ourpackage}
}
\description{
\packageDescription{ourpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ourpackage}
\packageIndices{ourpackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ourpackage}

Maintainer: \packageMaintainer{ourpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
