% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_patterns.R
\name{quadrat_patterns}
\alias{quadrat_patterns}
\title{Splits a point pattern into quadrats}
\usage{
quadrat_patterns(patterns)
}
\arguments{
\item{patterns}{A list of point patterns}
}
\value{
A list consisting of lists of the split point patterns
}
\description{
Splits a point pattern into quadrats
}
\examples{
pats <- list(spatstat.random::rpoispp(lambda = 200),
          spatstat.random::rpoispp(lambda = 100))
split_pats <- quadrat_patterns(pats)
}
