% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph is a data set, which contains three columns of data}
\format{A data frame with 3 variables:
\describe{
  \item{v1}{the first node in a link}
  \item{v2}{the second node in a link}
  \item{w}{the weight of an edge between the linked nodes}
}}
\usage{
wiki_graph
}
\description{
Wiki_graph is a data set, which contains three columns of data
}
\references{
\url{https://sv.wikipedia.org/wiki/Dijkstras_algoritm}
}
\keyword{datasets}
