% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.r
\name{dijkstra}
\alias{dijkstra}
\title{determine the shortest path to all the nodes, which are included in the network}
\usage{
dijkstra(wiki_graph, initial_node)
}
\arguments{
\item{wiki_graph}{a data frame, which contains three parameters: the connection path between two nodes and weight function}

\item{initial_node}{from which node, the trip begins}
}
\value{
the expected shortest journey, towards each and every node, through the entire network
}
\description{
determine the shortest path to all the nodes, which are included in the network
}
\examples{
wiki_graph <- data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6), v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5), w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)) dijkstra(wiki_graph, 1)

}
