% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph is a data set, which contains three row data}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
data(dijkstra)
}
\description{
Wiki_graph is a data set, which contains three row data
}
\examples{
data(wiki_graph)
wiki_graph is a data frame, which contains three row vectors
v1 and v2, signifies the links
w is a weight function


}
\references{
\url{https://sv.wikipedia.org/wiki/Dijkstras_algoritm}
}
\keyword{datasets}
